requirements.txt
Flask==1.1.1
json5==0.8.5
jsonschema==3.2.0
pandas==1.0.1
matplotlib==3.1.2
mysql-connector-python==8.0.17
numpy==1.18.1
requests==2.22.0
scikit-learn==0.22.1
seaborn==0.9.0
pika==1.1.0
prettytable==0.7.2
TPOT==0.11.1